+++ certbot-nginx/certbot_nginx/tests/parser_test.py
@@ -10,7 +10,7 @@ from certbot import errors
 from certbot.compat import os
 
 from certbot_nginx._internal import nginxparser
-from certbot_nginx import obj
+from certbot_nginx._internal import obj
 from certbot_nginx import parser
 from certbot_nginx.tests import util
